Login y Control de Login

+++ Controllers/TableroController.cs
@@ -18,6 +18,10 @@ namespace tl2_tp10_2023_NicoMagro.Controllers
 
         public IActionResult Index()
         {
+            if (!Logueado())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
             List<Tablero> tableros = repository.GetAll();
 
             if (tableros != null)
@@ -33,12 +37,20 @@ namespace tl2_tp10_2023_NicoMagro.Controllers
         [HttpGet]
         public IActionResult CrearTablero()
         {
+            if (!Logueado())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
             return View(new Tablero());
         }
 
         [HttpPost]
         public IActionResult CrearTablero(Tablero tablero)
         {
+            if (!Logueado())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
             repository.Create(tablero);
             return RedirectToAction("Index");
         }
@@ -46,12 +58,20 @@ namespace tl2_tp10_2023_NicoMagro.Controllers
         [HttpGet]
         public IActionResult EditarTablero(int id)
         {
+            if (!Logueado())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
             return View(repository.GetById(id));
         }
 
         [HttpPost]
         public IActionResult EditarTablero(Tablero tablero)
         {
+            if (!Logueado())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
             var tablero2 = repository.GetById(tablero.Id);
             tablero2.Nombre = tablero.Nombre;
             tablero2.Descripcion = tablero.Descripcion;
@@ -63,11 +83,24 @@ namespace tl2_tp10_2023_NicoMagro.Controllers
 
         public IActionResult EliminarTablero(int Id)
         {
+            if (!Logueado())
+            {
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+            }
             repository.Remove(Id);
 
             return RedirectToAction("Index");
         }
 
+        public bool Logueado()
+        {
+            return HttpContext.Session.Keys.Any();
+        }
+
+        private bool esAdmin()
+        {
+            return HttpContext.Session.Keys.Any() && ((int)HttpContext.Session.GetInt32("rol") == 1);
+        }
 
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
